bump globby
`globby` was a major release because of https://github.com/isaacs/node-glob/commit/0ffab1f3cbe2f3c9ec6eab21dc16c8e0d1fe3e0a, but that's IMHO not a breaking change and was already documented not to work. So I'm going to treat it as a bug fix here.

fixes #57